Add base layer component breakdown and start tests
This is an overly bloated commit because I went rogue and just started
writing a bunch of code without marking points of progress. The
following commits are going to be a large clean up effort for all the
code being added here, as well as unit tests to ensure correctness.